requirements.txt
-i https://pypi.org/simple
certifi==2018.11.29
chardet==3.0.4
idna==2.8
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
entrypoints==0.3
flake8==3.7.6
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.0
svgwrite==1.2.1